feat(web): SSR reform pages for SEO recovery + sitemap index
The reform detail page (/cambios/reforma/) was 100% client-side (fetch +
innerHTML), so Google saw an empty shell and the ~44k reform pages were
noindex'd on 2026-05-04. This restores their indexability:

- Add @astrojs/cloudflare adapter (output stays "static"; only this route
  opts into on-demand rendering via `export const prerender = false`).
- Rewrite cambios/reforma/index.astro to fetch the reform (+ omnibus/diff
  fallbacks) server-side and render the word-diff into the response HTML
  instead of lazily building it client-side. Unique <title>/description per
  reform, canonical without from/topic, noindex removed on the normal case
  (kept only for the 404/error state), 90-day Cache-Control.
- Split sitemap.xml into an index (sitemap-leyes.xml + sitemap-reformas.xml)
  so the ~44k reform URLs are finally listed for crawling.
- Add packages/web/wrangler.jsonc (compatibility_date + nodejs_compat).

KNOWN GAP, needs sign-off before this reaches prod: @astrojs/cloudflare v13+
(required by our Astro 6.1.9) dropped Cloudflare Pages support — it only
targets Workers with static assets (`wrangler deploy`). deploy.yml still
runs `wrangler pages deploy dist`, which will NOT execute the new on-demand
route (Pages would just 404 it). Switching requires a one-time manual
Cloudflare Dashboard step (move the leyabierta.es custom domain from the
Pages project to a Workers script) — flagged inline in deploy.yml and
wrangler.jsonc, not resolved silently here.
