+++ packages/web/package.json
@@ -8,6 +8,7 @@
 		"preview": "astro preview"
 	},
 	"dependencies": {
+		"@astrojs/cloudflare": "^14.1.3",
 		"@astrojs/react": "^5.0.3",
 		"@lottiefiles/dotlottie-react": "^0.19.2",
 		"astro": "6.1.9",
